changed divide function to return double

+++ src/main/java/com/example/springdocker/MathCalc.java
@@ -10,7 +10,7 @@ public class MathCalc {
         return a * b;
     }
 
-    public float divide(float a, float b) {
+    public double divide(double a, double b) {
         return a / b;
     }
 }
